pyproject.toml
[tool.poetry]
name = "sage-dashboard"
version = "0.5.0"
description = "SAGE (Sex-Aware Genomics Explorer) - Illuminating the gaps in sex-stratified genomics data"
authors = ["Lina Faller <lina@linafaller.com>"]
readme = "README.md"
license = "MIT"
packages = [{include = "sage", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.28.0"
pandas = "^2.0.0"
plotly = "^5.18.0"
supabase = "^2.27.1"
geoparse = "^2.0.4"
biopython = "^1.84"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.0.0"
ruff = "^0.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=src --cov-report=term-missing --cov-report=html"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
